Prompt Containers
Currently we've created a shared component that will allowed to be used
throughout different routes and utilizing the props passed to determine
state.

In the PromptContainer you will notice there are a few new methods being
used. getInitialState is parts of the framework and is used to set
inital state. The other method `updateUser` does just what it says, and
is fired on the change of the input text field.